Add files via upload
The intended audience of this material is to those have an idea to learn Bayesian Inference. This work is not an attempt to replace a textbook but rather provides a road map for the journey with the Bayesian perspective of Statistical Inference. However, the presentation does attempt to compile in an ordered way but not as typically followed in a textbook, in general. The material is an outcome of a Bayesian Introductory course from a teaching-learning perspective. Hence, this material covers Student’s projects as illustrative examples based on simple models.